requirements.txt
# Unified Trading Engine — dependencies
alpaca-py>=0.33.1
python-dotenv>=1.0.0
pyyaml>=6.0
requests>=2.28.0
pytz>=2023.3
pandas>=2.0.0
numpy>=1.24.0
